create: ArtifactLoader class.
This class is manager of Contract SourceCode, bytecodes, sourceMap.
